pulled out getproviderfrommodel
