+++ shared/utils/index.ts
@@ -1,3 +1,4 @@
 export * from "./logger";
 export * from "./googleAuth";
+export * from "./llmProvider";
 // Export any other utilities you have
