+++ workers/agent-service/src/tools/utils.ts
@@ -78,15 +78,3 @@ export function getOpenAIContent(screenshots: EnhancedScreenshotResult[]) {
 
   return content;
 }
-
-export function getProviderFromModel(model: string): LLMProvider {
-  if (model.includes("gpt")) {
-    return "openai";
-  } else if (model.includes("gemini")) {
-    return "vertex-ai";
-  } else if (model.includes("llama")) {
-    return "groq";
-  } else {
-    throw new Error(`Unsupported model: ${model}`);
-  }
-}
